refactor(#130): auth 훅 분리
